More frontend stats & switched to aspnet.razor as the view engine for frontend.

+++ src/CoiniumServ/Daemon/Responses/MiningInfo.cs
@@ -20,6 +20,9 @@
 //     license or white-label it as set out in licenses/commercial.txt.
 // 
 #endregion
+
+using System;
+
 namespace Coinium.Daemon.Responses
 {
     public class MiningInfo
@@ -32,7 +35,7 @@ namespace Coinium.Daemon.Responses
         public bool Generate { get; set; }
         public int GenProcLimit { get; set; }
         public int HashesPerSec { get; set; }
-        public int NetworkHashps { get; set; }
+        public UInt64 NetworkHashps { get; set; }
         public int PooledTx { get; set; }
         public bool Testnet { get; set; }
     }
